+++ client_dev.py
@@ -1,0 +1,7 @@
+"""
+Simple shim for running the client program during development.
+"""
+import ga4gh.cli
+
+if __name__ == "__main__":
+    ga4gh.cli.client_main()
