--- ga4gh/scripts/__init__.py
@@ -1,3 +1,0 @@
-"""
-Package holding code for executable scripts.
-"""
